# Optimize show(io::IO, m::Module)