@Sacha0 2021-10-29
Tagged for backport to 1.6 and 1.7 as a low-risk optimization :).