@gustaphe 2021-11-16
Why not `join(io, nameof(module), ".")`?