@KristofferC 2021-11-16
> Why not `join(io, fullname(module), ".")`?

The PR here was for performance as shown in the first post. Did you benchmark that version and find that it performs the same as the one proposed here?