@gustaphe 2021-11-19
My bad, `fullname` of course allocates. I didn't realize this recursive method didn't use the fullname tuple.

(the `join(io...)` version is inbetween the two methods in the PR for speed)